requirements_colab.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.50.0,<5.0.0
tokenizers>=0.21.0,<0.22.0
peft>=0.6.0
accelerate>=0.24.0
datasets>=2.14.0
evaluate>=0.4.0
open-clip-torch>=2.26.0
ultralytics>=8.3.0
groundingdino-hf

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.0.0
scikit-learn>=1.3.0
opencv-python>=4.7.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
requests>=2.31.0
gdown>=4.7.0
huggingface-hub>=0.17.0

# Monitoring and notebooks
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipywidgets>=8.0.0
google-colab>=1.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
